Fix bug in incremental resource merge in case of styleable.
The merge blob writing code didn't account for ignored resources.
This is the case of attr resources declared inside a declare-styleable
resource which are not merged. The blog writting code didn't check
for this and wrote this as new resources, moving them out of the
declare-styleable

Change-Id: Ia781eeedd06c1323dac7593c192b172ee4e6fc04
(cherry picked from commit cbf300f95b1dbb18f7b2adbb50cc986da6d50037)