Duplicate the VariantManager to prepare for the Java conversion.
This deals with the fact that the conversion would be too heavy
for git to know the file was renamed. This way, the next
CL can truly show the file diff and it's easier to review.

Change-Id: I28cebc9f8133ab7956c8d703a6c18ba993ef82c0
(cherry picked from commit 2c437d6b45bf4530e1fdb8196c4a6f9d9e49a5ee)