Specify package name in build.gradle by default
When the package name is specified in the build.gradle file, the
manifest file's package element is only used to control where the R
class is and to resolve package relative activity/service
registrations. This has the nice property of separating the
application id (the package name specified in build.gradle) from the
implementation package (used in the code), which helps clean up for
example the refactoring behavior. (It currently skips the package
attribute in the manifest, since that could lead to accidentally
changing the application id, and a number of users have asked about
this.)

Change-Id: I8a420a9a3e1981b64311d17287b8792393e99d6a
(cherry picked from commit edb65682ab19f86b7d43996d4c0ac29028278441)