Add a lint check which checks whether Gradle libraries are up to date
We already have a check which compares your version numbers against
the available versions in a local Gradle repository (for the Android
support artifacts for example) or against a set of known version
numbers for a couple of libraries.

This CL adds a new lint check, off by default, which looks up the
version numbers on MavenCentral, and if a newer version is found,
lists the newer version. This can be used occasionally to check to
see if there are opportunities to upgrade your dependencies.

69694: Lint "Obsolete Gradle Dependency"

Change-Id: I3ad7dfbb08b871d16200db5ece41bd94a8bca2b9
(cherry picked from commit 73bdf85189f2e362e908c1d94e438393aa482367)