Handle BOM (byte order marks) in UTF encoded XML files
We recently tightened the requirement to use UTF instead of default system
encodings in XML files. However, the XML parsing code did not handle the
optional byte order mark at the beginning of UTF-8 files. This fixes that,
and consolidates the various local implementations of XML parsing with calls
to a central utility.

See issue http://b.android.com/69884

Change-Id: I1955de31d4f4540214f177c5a311f73c9bf383a1

Conflicts:
	gradle-import/src/main/java/com/android/tools/gradle/eclipse/GradleImport.java