SDK loader: mark build-tools as broken if they don't match the current OS. DO NOT MERGE.
This loads the build-tools as "broken packages" so they are
visible in the SDK Manager but they are not functional as
actual build-tools.

This is only useful for users who manually create build-tools
local packages -- when installed via the SDK Manager, the OS
is checked before download so it's not technically possible for
a user to download and install an invalid build-tools package.

(cherry picked from commit a4a587c3910e7c20bc4ff753b11f079fc77d6ac9
and back-ported to the older Tools 22.6 API for Archive.Os)

Change-Id: Ib55c12c16e4838c0140717026c738a9f736625cc