Treat uses-sdk in overlay manifests differently.
We used to treat uses-sdk declaration in overlay manifests the same way as
when they are declared in main or library manifests. However, most overlay
manifests are just declaring a few elements and uses-sdk is most of the time
not part of them. This result in implicit declaration (with a minSdk version
set to 1) being added to the overlay manifest.

Later when trying to merge library manifest with higher minSdk version, the
merger would fail.

With this change, the implicit uses-sdk versions and associated permissions
are not added to an overlay when there is no uses-sdk element declared
in that overlay manifest.

Change-Id: Ia200f44c2ef64524a07f01e54559b8e81cabbc6e
(cherry picked from commit 7b11f04b5caea18f7ca83c6c3185a135ef37a03b)