Bind ${applicationId} when none specified in build.gradle
The manifest merger was expecting the packageName/applicationId to always
be provided by the gradle plugin but it is not.
In that case, extract the main manifest file package name and automatically
binds it for packageName and applicationId placeholder replacement.

Change-Id: Idad54501e16437d96d3a0447d03dac19a4b1328b
(cherry picked from commit fb5f42af6cda119dd0a2345b9b646362f9d2c917)