Fix multi-apk logic to use versionCode.
The store logic does not actually look for a best
match, but instead look for all compatible devices
and then simply select the one with the highest versionCode.

Changed the model to add the versionCode to the SplitOutput
so that the logic can use it.

Change-Id: I15717a4261b6697eb491cf51679d4dbe5ad57806
(cherry picked from commit 6e986752c1f9696ad81658eace4381d758b5e7af)