Always copy the default config when building the variant config.
This is important so that in case of no flavors, the 2 variants have
different versions of the mergedFlavor (which is initialized with
the defaultConfig).

Otherwise, editing a mergedFlavor on a variant will actually impact
both variants.

Bug: 68528

Change-Id: I07e7bcf578a41a81d9c877b19fe482595293b960
(cherry picked from commit 75d68c764c60b6a9d21a7743268ca725b6acc258)